review(snes): address PR #2942 review comments
- Fix stale snes_tests_cpu/snes_tests_spc module-path references left in
  manifest.json notes after the gilyon_cpu_tests/gilyon_spc_tests rename.
- Reduce tick budget from 4B to the established 400M (matches
  blargg_apu_tests.rs convention); both ROMs only need ~87M ticks to reach
  frame 2000, so a hung/regressed run now fails faster in CI.
- Extract assert_rom_screen_crc into rom_runner.rs to remove the
  read-ROM/run-oracle/assert duplication between gilyon_cpu_tests.rs and
  gilyon_spc_tests.rs.
- Clarify in a code comment that PLB's emulation-mode read8 call already
  goes through the same ticked tick_read path pull8 itself uses, so there's
  no cycle-accuracy divergence.
