+++ src/snes/integration_tests/gilyon_spc_tests.rs
@@ -1,6 +1,4 @@
-use super::rom_runner::{RunConfig, RunExitReason, RunOracle, run_rom_with_oracle};
-use std::fs;
-use std::path::Path;
+use super::rom_runner::{RunConfig, assert_rom_screen_crc};
 
 const GILYON_TESTS_SPC_ROOT: &str = "roms/snes/automated_tests/gilyon_tests/spctest";
 
@@ -10,26 +8,16 @@ mod tests {
 
     #[test]
     fn spctest_passes_all_1368_tests() {
-        // Reaches "Success" at test 0557 (hex), the last of 1368 SPC-700 tests.
-        let path = Path::new(GILYON_TESTS_SPC_ROOT).join("spctest.sfc");
-        let rom = fs::read(&path)
-            .unwrap_or_else(|err| panic!("failed to read ROM {}: {err}", path.display()));
-        let result = run_rom_with_oracle(
-            &rom,
+        // Reaches "Success" at test 0557 (hex), the last of 1368 SPC-700
+        // tests. Needs ~87M ticks to reach frame 2000; 400M (the same
+        // budget used throughout blargg_apu_tests.rs) gives comfortable
+        // headroom without letting a hung/regressed run grind for long.
+        assert_rom_screen_crc(
+            GILYON_TESTS_SPC_ROOT,
             "spctest.sfc",
-            RunConfig::new(4_000_000_000, 0),
-            RunOracle::ScreenCrc {
-                frames: 2000,
-                expected_crc: 0x87CD_986B,
-            },
-        );
-        assert!(
-            result.passed && result.exit_reason == RunExitReason::ScreenCrcFrame,
-            "spctest.sfc: expected screen-CRC PASS at frame 2000, \
-             got crc=0x{:08X} passed={} exit={:?}",
-            result.screen_crc32,
-            result.passed,
-            result.exit_reason
+            2000,
+            0x87CD_986B,
+            RunConfig::new(400_000_000, 0),
         );
     }
 }
